pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-optimize-ci"
version = "0.1.0"
description = "Cost-quality-latency optimizer and CI gate for RAG QA systems"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Harmeet Singh" }]
dependencies = []

[project.optional-dependencies]
full = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7"
]
dev = [
  "pytest>=8.0,<9"
]

[project.scripts]
ragopt = "ragopt.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["ragopt*"]
